Sun, Jul 2, 2023, 8:02 PM +03:00

+++ sentimental-cash/cash.py
@@ -1,6 +1,6 @@
-import m
+
 def main():
-    cents = float(input())
+    cents = int(input())
 
 
     quarters = calculate_quarters(cents)
